+++ examples/matplotlib/mpl_plot_compare.py
@@ -3,6 +3,7 @@ Compare Plot
 ============
 
 _thumb: .5, .5
+_example_title: Comparison plot
 """
 import matplotlib.pyplot as plt