+++ examples/matplotlib/mpl_plot_density.py
@@ -3,6 +3,7 @@ Density Plot
 ============
 
 _thumb: .5, .5
+_example_title: Plot density
 """
 import matplotlib.pyplot as plt