+++ examples/matplotlib/mpl_plot_pair_kde.py
@@ -3,6 +3,7 @@ KDE Pair Plot
 =============
 
 _thumb: .2, .5
+_example_title: KDE pair plot
 """
 import matplotlib.pyplot as plt