+++ examples/matplotlib/mpl_plot_pair_point_estimate.py
@@ -3,6 +3,7 @@ Point Estimate Pairplot
 =======================
 
 _thumb: .2, .5
+_example_title: Pair plot with point estimate markings
 """
 import matplotlib.pyplot as plt