+++ examples/matplotlib/mpl_plot_parallel_minmax.py
@@ -3,6 +3,7 @@ MinMax Parallel Plot
 ====================
 
 _thumb: .2, .5
+_example_title: Parallel plot with MinMax normalization
 """
 import matplotlib.pyplot as plt