+++ examples/matplotlib/mpl_plot_trace.py
@@ -3,6 +3,7 @@ Traceplot
 =========
 
 _thumb: .1, .8
+_example_title: Plot trace
 """
 import matplotlib.pyplot as plt